+++ src/main/java/com/imsweb/algorithms/surgery/SiteSpecificSurgeryUtils.java
@@ -161,8 +161,8 @@ public final class SiteSpecificSurgeryUtils {
 
                 rowDto.setCode(row.getCode());
                 rowDto.setDescription(row.getDescription());
-                rowDto.setLevel(row.getLevel());
-                rowDto.setLineBreak(row.isBreak());
+                rowDto.setLevel(row.getLevel() == null ? 0 : row.getLevel());
+                rowDto.setLineBreak(row.isBreak() == null ? Boolean.FALSE : row.isBreak());
 
                 rows.add(rowDto);
             }